requirements.txt
numpy
pandas
scikit-learn
gensim
pyyaml
seaborn
matplotlib